Multiplication result converted to larger type
CWE-190
CWE-192
CWE-197
CWE-681

To fix this safely, force the multiplication to be performed in size_t from
the start, and use that computed size for malloc.

This preserves functionality while preventing overflow in unsigned int
intermediate arithmetic.

No new headers or dependencies are required.

+++ src/BarShader.cpp
@@ -170,7 +170,9 @@ void CBarShader::Draw( wxDC* dc, int iLeft, int iTop, bool bFlat )
 	}
 
 	// Render the bar into a raw buffer
-	unsigned char * buf = (unsigned char *) malloc(m_Width * m_Height * 3);
+	unsigned char * buf = (unsigned char *) malloc(
+		static_cast<size_t>(m_Width) *
+		static_cast<size_t>(m_Height) * 3u);
 
 	if (bFlat) {
 		// draw flat bar
